requirements.txt
requests
python-dotenv
plotly
pandas
finnhub-python